+++ src/Tesseract.java
@@ -1,15 +1,15 @@
 public class Tesseract extends Cube {
-    int d;
-    Tesseract(int a, int b, int c, int d) {
+    double d;
+    Tesseract(double a, double b, double c, double d) {
         super(a, b, c);
         this.d = d;
     }
 
-    int getPerimeter() {
+    double getPerimeter() {
         return this.a + this.b + this.c + this.d;
     }
 
-    int getArea() {
+    double getArea() {
         return this.a * this.b * this.c * this.d;
     }
 }
